feat: add shadcn dependency and integrate styles into globals css
