+++ frontend/package.json
@@ -70,6 +70,7 @@
     "react-day-picker": "^10.0.0",
     "react-dom": "19.2.5",
     "recharts": "3.8.0",
+    "shadcn": "^4.12.0",
     "sonner": "^2.0.7",
     "streamdown": "^2.5.0",
     "tailwind-merge": "^3.5.0",
